requirements.txt
torch==1.8.0
timm==0.8.3.dev0
scikit-image
ptflops
easydict
PyYAML
pillow
torchvision==0.9.0
opencv-python
